Add Cajita dependency on core; Move ParameterPack to core
