+++ cajita/src/Cajita_Halo.hpp
@@ -14,7 +14,8 @@
 
 #include <Cajita_Array.hpp>
 #include <Cajita_IndexSpace.hpp>
-#include <Cajita_ParameterPack.hpp>
+
+#include <Cabana_Core.hpp>
 
 #include <Kokkos_Core.hpp>
 
@@ -688,11 +689,12 @@ class Halo
                const Kokkos::View<int**, memory_space>& steering,
                const int element_idx,
                const std::integral_constant<std::size_t, 0>,
-               const ParameterPack<ArrayViews...>& array_views )
+               const Cabana::ParameterPack<ArrayViews...>& array_views )
     {
         // If the pack element_idx is in the current array, pack it.
         if ( 0 == steering( element_idx, 1 ) )
-            packElement( buffer, steering, element_idx, get<0>( array_views ) );
+            packElement( buffer, steering, element_idx,
+                         Cabana::get<0>( array_views ) );
     }
 
     // Pack an array into a buffer.
@@ -702,12 +704,13 @@ class Halo
                const Kokkos::View<int**, memory_space>& steering,
                const int element_idx,
                const std::integral_constant<std::size_t, N>,
-               const ParameterPack<ArrayViews...>& array_views )
+               const Cabana::ParameterPack<ArrayViews...>& array_views )
     {
         // If the pack element_idx is in the current array, pack it.
         if ( N == steering( element_idx, 1 ) )
         {
-            packElement( buffer, steering, element_idx, get<N>( array_views ) );
+            packElement( buffer, steering, element_idx,
+                         Cabana::get<N>( array_views ) );
             return;
         }
 
@@ -723,7 +726,7 @@ class Halo
                      const Kokkos::View<int**, memory_space>& steering,
                      ArrayViews... array_views ) const
     {
-        auto pp = makeParameterPack( array_views... );
+        auto pp = Cabana::makeParameterPack( array_views... );
         Kokkos::parallel_for(
             "pack_buffer",
             Kokkos::RangePolicy<ExecutionSpace>( exec_space, 0,
@@ -823,12 +826,12 @@ class Halo
                  const Kokkos::View<int**, memory_space>& steering,
                  const int element_idx,
                  const std::integral_constant<std::size_t, 0>,
-                 const ParameterPack<ArrayViews...>& array_views )
+                 const Cabana::ParameterPack<ArrayViews...>& array_views )
     {
         // If the unpack element_idx is in the current array, unpack it.
         if ( 0 == steering( element_idx, 1 ) )
             unpackElement( reduce_op, buffer, steering, element_idx,
-                           get<0>( array_views ) );
+                           Cabana::get<0>( array_views ) );
     }
 
     // Unpack an array from a buffer.
@@ -839,13 +842,13 @@ class Halo
                  const Kokkos::View<int**, memory_space>& steering,
                  const int element_idx,
                  const std::integral_constant<std::size_t, N>,
-                 const ParameterPack<ArrayViews...>& array_views )
+                 const Cabana::ParameterPack<ArrayViews...>& array_views )
     {
         // If the unpack element_idx is in the current array, unpack it.
         if ( N == steering( element_idx, 1 ) )
         {
             unpackElement( reduce_op, buffer, steering, element_idx,
-                           get<N>( array_views ) );
+                           Cabana::get<N>( array_views ) );
             return;
         }
 
@@ -863,7 +866,7 @@ class Halo
                        const Kokkos::View<int**, memory_space>& steering,
                        ArrayViews... array_views ) const
     {
-        auto pp = makeParameterPack( array_views... );
+        auto pp = Cabana::makeParameterPack( array_views... );
         Kokkos::parallel_for(
             "unpack_buffer",
             Kokkos::RangePolicy<ExecutionSpace>( exec_space, 0,
