+++ core/src/CMakeLists.txt
@@ -20,6 +20,7 @@ set(HEADERS_PUBLIC
   Cabana_MemberTypes.hpp
   Cabana_NeighborList.hpp
   Cabana_Parallel.hpp
+  Cabana_ParameterPack.hpp
   Cabana_Slice.hpp
   Cabana_SoA.hpp
   Cabana_Sort.hpp
