+++ core/src/Cabana_Core.hpp
@@ -20,6 +20,7 @@
 #include <Cabana_MemberTypes.hpp>
 #include <Cabana_NeighborList.hpp>
 #include <Cabana_Parallel.hpp>
+#include <Cabana_ParameterPack.hpp>
 #include <Cabana_Slice.hpp>
 #include <Cabana_SoA.hpp>
 #include <Cabana_Sort.hpp>
